requirementsOld.txt
Flask==2.3.2
openai==1.78.1
google-cloud-texttospeech==2.14.1
python-dotenv==1.0.1
flask-cors==3.0.10